chore(students): move codigo para students app
