+++ src/services/api/models.py
@@ -3,15 +3,7 @@ import uuid
 from django.db import models
 from django.core.validators import MaxValueValidator, MinValueValidator
 
-class Student(models.Model):
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    name = models.CharField(max_length=128)
-    email = models.EmailField(max_length=128, unique=True)
-    usp_number = models.CharField(max_length=8, unique=True, null=True, blank=True)
-    code = models.CharField(max_length=4, unique=True, null=True, blank=True)
-
-    def __str__(self) -> str:
-        return f"Student: {self.name}"
+from students.models import Student
 
 class Speaker(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
