+++ src/services/students/models.py
@@ -1,3 +1,13 @@
 from django.db import models
+import uuid
 
 # Create your models here.
+class Student(models.Model):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    name = models.CharField(max_length=128)
+    email = models.EmailField(max_length=128, unique=True)
+    usp_number = models.CharField(max_length=8, unique=True, null=True, blank=True)
+    code = models.CharField(max_length=4, unique=True, null=True, blank=True)
+
+    def __str__(self) -> str:
+        return f"Student: {self.name}"
