+++ src/services/students/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Student
+
+
+class StudentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Student
+        fields = ['id', 'name', 'email', 'usp_number']
