+++ src/services/students/tests.py
@@ -1,3 +1,3 @@
 from django.test import TestCase
 
-# Create your tests here.
+# TODO: criar testes para as views aqui
