+++ tests/test_studentLogin.py
@@ -1,11 +1,11 @@
 from rest_framework.test import APITestCase, APIClient
 from rest_framework.views import APIView
 from django.urls import reverse
-from api.models import Student
+from students.models import Student
 from rest_framework import status
 
 from unittest.mock import patch
-from api.views import StudentLogin
+from students.views import StudentLogin
 
 class StudentLoginViewTest(APITestCase):
     def setUp(self):
